feat: :sparkles: added import method and update command to accept array of datas

+++ b3-challenge/database/migrations/2023_05_03_142805_create_b3_table.php
@@ -18,9 +18,9 @@ return new class extends Migration
             $table->string('isin');
             $table->string('asst');
             $table->integer('balqty');
-            $table->float('tradavrgpric');
+            $table->double('tradavrgpric');
             $table->integer('pricfctr');
-            $table->float('balval');
+            $table->double('balval');
             $table->timestamps();
         });
     }
